Switch to log package for safety, logging improvements, output committer
fmt.Fprintf is not concurrent safe, so switched over to log which uses
locking on the inside to prevent overlapping outputs.

The output is now also committed in a separate step after the the final
attempt succeeds (or if it fails, no final output is saved). For
example, if a command id=3 failed on 3 different hosts, and then on the
4th attempt succeeds, it would generate output files cmd_3-attempt0.log
to cmd_3-attempt2.log, and then cmd_3-attempt4.log will have been
overwritten to cmd_3-final.log. This is inspired by OutputCommitter
behavior from Hadoop.